refactor: split PhotoGallery into components
