+++ src/modules/CountryPage/components/CountryPageView/components/PhotoGallery/PhotoGallery.tsx
@@ -3,23 +3,14 @@ import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import styled from 'styled-components';
-import { useTranslation } from 'react-i18next';
-import { useDispatch, useSelector } from 'react-redux';
-import { database } from 'services';
-import { loadCountry, loadUserInfo } from 'store/actions';
 
 import IconButton from '@material-ui/core/IconButton';
 import FullscreenIcon from '@material-ui/icons/Fullscreen';
 import FullscreenExitIcon from '@material-ui/icons/FullscreenExit';
-import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos';
-import ArrowForwardIosIcon from '@material-ui/icons/ArrowForwardIos';
-import Grid from '@material-ui/core/Grid';
-import Rating from '@material-ui/lab/Rating';
-import Chip from '@material-ui/core/Chip';
-import Badge from '@material-ui/core/Badge';
 
-import { LanguagesType, SliderDataType, State } from 'types';
+import { LanguagesType, SliderDataType } from 'types';
 import classes from './PhotoGallery.module.scss';
+import { PrevArrow, NextArrow, Slide } from './components';
 
 type PhotoGalleryProps = {
   sliderData: SliderDataType[];
@@ -42,47 +33,8 @@ const SliderContainer = styled('div')`
   }
 `;
 
-function NextArrow({
-  onClick,
-}: {
-  onClick: React.MouseEventHandler<HTMLButtonElement>;
-}) {
-  return (
-    <button
-      type="button"
-      className={`${classes.slickArrow} ${classes.slickNextArrow}`}
-      onClick={onClick}
-    >
-      <ArrowForwardIosIcon fontSize="large" />
-    </button>
-  );
-}
-
-function PrevArrow({
-  onClick,
-}: {
-  onClick: React.MouseEventHandler<HTMLButtonElement>;
-}) {
-  return (
-    <button
-      type="button"
-      className={`${classes.slickArrow} ${classes.slickPrevArrow}`}
-      onClick={onClick}
-    >
-      <ArrowBackIosIcon fontSize="large" />
-    </button>
-  );
-}
-
 export const PhotoGallery: FC<PhotoGalleryProps> = ({ sliderData, lang }) => {
   const [fullScreen, setFullScreen] = useState(false);
-  const { t } = useTranslation();
-  const countryId = useSelector((state: State) => state.country.id);
-  const userId = useSelector((state: State) => state.userInfo.id);
-  const userRates = useSelector(
-    (state: State) => state.userInfo.attractionRates
-  );
-  const dispatch = useDispatch();
 
   const settings = {
     customPaging(i: number) {
@@ -133,86 +85,10 @@ export const PhotoGallery: FC<PhotoGalleryProps> = ({ sliderData, lang }) => {
     };
   }, []);
 
-  const onRatingChange = (
-    evt: React.ChangeEvent<HTMLInputElement>,
-    value: number | null
-  ) => {
-    database
-      .setRating(countryId, evt.target.name, userId, value as number)
-      .then((res) => {
-        dispatch(loadCountry(countryId));
-        dispatch(loadUserInfo(userId));
-      })
-      .catch((err) => err);
-  };
-
   const slides = sliderData
-    ? sliderData.map((slide) => {
-        const userRate = userRates.find((rate) => rate.attrId === slide.id);
-        return (
-          <Grid container className={classes.slickImage} key={slide.id}>
-            <Grid item md={12} lg={7}>
-              <img
-                src={slide.photo}
-                alt={slide.name[lang as keyof LanguagesType]}
-              />
-            </Grid>
-            <Grid item md={12} lg={5} className={classes.slickImageInfo}>
-              <div className={classes.slickImageInfoWrapper}>
-                <div className={classes.slickImageCaption}>
-                  <p className={classes.slickImageTitle}>
-                    {slide.name[lang as keyof LanguagesType]}
-                  </p>
-                  <p className={classes.slickImageDescription}>
-                    {slide.description[lang as keyof LanguagesType]}
-                  </p>
-                </div>
-                <div className={classes.slickAttractionRating}>
-                  {userId && (
-                    <div>
-                      <p className={classes.slickAttractionRatingHeader}>
-                        {t('Rate This Place')}
-                      </p>
-                      <Rating
-                        name={slide.id}
-                        value={userRate ? userRate.rating : 0}
-                        precision={0.5}
-                        max={5}
-                        size="large"
-                        className={classes.slickAttractionRatingStars}
-                        onChange={(evt, value) =>
-                          onRatingChange(
-                            evt as React.ChangeEvent<HTMLInputElement>,
-                            value
-                          )
-                        }
-                      />
-                    </div>
-                  )}
-                  <div className={classes.slickAttractionRatingStats}>
-                    <Badge
-                      badgeContent={
-                        slide.rating.count
-                          ? slide.rating.sum / slide.rating.count
-                          : '0'
-                      }
-                      color="secondary"
-                    >
-                      <Chip label={t('Rating')} />
-                    </Badge>
-                    <Badge
-                      badgeContent={slide.rating.count || '0'}
-                      color="primary"
-                    >
-                      <Chip label={t('Voted')} />
-                    </Badge>
-                  </div>
-                </div>
-              </div>
-            </Grid>
-          </Grid>
-        );
-      })
+    ? sliderData.map((slideData) => (
+        <Slide key={slideData.id} slideData={slideData} lang={lang} />
+      ))
     : null;
 
   const sliderRef = useRef<HTMLDivElement>(null);
