+++ src/modules/CountryPage/components/CountryPageView/components/PhotoGallery/components/index.ts
@@ -1,0 +1,3 @@
+export { NextArrow } from './NextArrow';
+export { PrevArrow } from './PrevArrow';
+export { Slide } from './Slide';
